Refactor animation and transition modules structure
Renamed `animations::utils` to `animations::core` and `transitions::utils` to `transitions::config` for clearer module organization. Updated all imports and references accordingly throughout the codebase. This improves code clarity and maintainability by better reflecting the purpose of these modules.
