+++ src/animations/mod.rs
@@ -1,6 +1,6 @@
 pub mod colors;
+pub mod core;
 pub mod platform;
 pub mod spring;
 pub mod transform;
 pub mod tween;
-pub mod utils;
