+++ src/transitions/mod.rs
@@ -1,2 +1,2 @@
+pub mod config;
 pub mod page_transitions;
-pub mod utils;
